backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
python-dotenv
psycopg2 